requirements.txt
discord.py
multidict
pytube
youtube-search
pynacl
ffmpeg
python-vlc
spotipy
# youtube_dl       pip install git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube_dl